requirements.txt
# 简化版依赖 - 只保留DeepSeek翻译分析所需
requests>=2.25.0
python-dotenv>=0.19.0
